Add validator for Bitcoin Cash CassAddr.
This validator only validates the nre CassAddr format. The legacy format
is the same as the BTC address, so IPED will classify and validate it as
a BTC address. The is no way off telling the difference (one may try to
look in the blockchain, but it will only work for used addresses).