Fix issue with HtmlViewer not finding page resources
As IPED is now storing subitens in containers, the file passed to
HtmlViewer is located in a temporary folder, and the resources in the
default case folder. As some parsers generate HTML with relative paths,
the viewer cannot find these resources anymore. Fix by editing the tags
in html when it is loaded, fixing the reference.