RegistryParser tests working* locally.
*need fix the language package

Also need help with the regripper program when trying to build on
github. Can it be done? For now i'm setting the regripper path locally
so it can be tested... What should I do in this case that we need an
external application?

Minor changes to all parsers tests so it could test supported types
method from parser.