Premade the WinXTimeLineParser (Need debug).
Right now, for some weird reason, as far as I debugged, the sqlite3
parser is not filling the table names. Without the table names the
winxtimelineparser throws an exception whenever it tries to look for a
specified table on the Select query.

For now I'll try to test parsers that use libmsiecf using the release
version.