UsnJrnlParserTest finally done.
A lot of debugging was needed. Thanks for the  tip @lfcnassif! Returning
the file solved one of the problems. Others showed up though...

Had to manually delete line at the journal bin file.

Nothing is done with the handler in this parser so could not retrieve
the handler content as String.