#998: Improve progress bar accuracy and ETA calculation
Refactors the progress bar logic in `ProgressFrame` to provide a
smoother and more accurate representation of the processing status.

The progress is now calculated as an average of both processed data
volume and the number of processed items. This prevents the progress
bar from appearing stuck when handling embeded items with no content
and items with models from UfedXmlReader.

Additionally, the estimated time remaining (ETA) is now calculated
using the maximum of the time left for volume and items, resulting
in a more realistic and reliable estimate.